pyproject.toml
[project]
name = "criticat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "FarDust", email = "gabriel.faundez@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi-mcp>=0.2.0",
    "langchain-google-vertexai>=2.0.19",
    "langchain>=0.3.23",
    "langgraph>=0.3.27",
    "mcp[cli]>=1.6.0",
    "pdf2image>=1.17.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.0.0",
    "typer>=0.15.2",
    "dependency-injector>=4.46.0",
]

[project.scripts]
criticat-cli = "criticat.interfaces.cli:app"
criticat-mcp = "criticat.interfaces.server:main" # Point to the main function
criticat-api = "criticat.interfaces.mcp:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.criticat]
default_joke_mode = "default"

[tool.criticat.server]
default_host = "0.0.0.0"
default_port = 8000

[dependency-groups]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
]
dev = [
    "fpdf2>=2.8.2",
]
